+++ src/main/java/com/larsvansoest/runelite/clueitems/EmoteClueItemsPlugin.java
@@ -32,10 +32,10 @@ import com.google.inject.Provides;
 import com.larsvansoest.runelite.clueitems.data.EmoteClueImage;
 import com.larsvansoest.runelite.clueitems.data.util.EmoteClueImages;
 import com.larsvansoest.runelite.clueitems.overlay.EmoteClueItemOverlay;
-import com.larsvansoest.runelite.clueitems.toolbar.EmoteClueItemsPanel;
-import com.larsvansoest.runelite.clueitems.toolbar.RequirementPanelProvider;
-import com.larsvansoest.runelite.clueitems.toolbar.component.EmoteClueItemsPanelPalette;
-import com.larsvansoest.runelite.clueitems.toolbar.progress.RequirementStatusManager;
+import com.larsvansoest.runelite.clueitems.ui.EmoteClueItemsPanel;
+import com.larsvansoest.runelite.clueitems.ui.requirement.RequirementPanelProvider;
+import com.larsvansoest.runelite.clueitems.ui.EmoteClueItemsPanelPalette;
+import com.larsvansoest.runelite.clueitems.progress.RequirementStatusManager;
 import javax.inject.Inject;
 import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.Client;
