+++ src/main/java/com/larsvansoest/runelite/clueitems/data/util/EmoteClueAssociations.java
@@ -30,8 +30,8 @@ package com.larsvansoest.runelite.clueitems.data.util;
 
 import com.larsvansoest.runelite.clueitems.data.EmoteClueDifficulty;
 import com.larsvansoest.runelite.clueitems.data.EmoteClueItem;
-import com.larsvansoest.runelite.clueitems.toolbar.RequirementPanelProvider;
-import com.larsvansoest.runelite.clueitems.toolbar.progress.RequirementStatusManager;
+import com.larsvansoest.runelite.clueitems.ui.requirement.RequirementPanelProvider;
+import com.larsvansoest.runelite.clueitems.progress.RequirementStatusManager;
 import com.larsvansoest.runelite.clueitems.vendor.runelite.client.plugins.cluescrolls.clues.EmoteClue;
 import java.util.AbstractMap;
 import java.util.Arrays;
