--- src/main/java/com/larsvansoest/runelite/clueitems/toolbar/component/requirement/foldable/FoldIcon.java
@@ -1,55 +1,0 @@
-/*
- * BSD 2-Clause License
- *
- * Copyright (c) 2020, Lars van Soest
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
- * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
- * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package com.larsvansoest.runelite.clueitems.toolbar.component.requirement.foldable;
-
-import com.larsvansoest.runelite.clueitems.data.EmoteClueImage;
-import javax.swing.ImageIcon;
-import javax.swing.JLabel;
-
-public class FoldIcon extends JLabel
-{
-	private static class Arrow {
-		static ImageIcon DOWN = new ImageIcon(EmoteClueImage.Toolbar.Chevron.DOWN);
-		static ImageIcon LEFT = new ImageIcon(EmoteClueImage.Toolbar.Chevron.LEFT);
-	}
-
-	public FoldIcon() {
-		super.setIcon(Arrow.LEFT);
-	}
-
-	public void unfold()
-	{
-		super.setIcon(Arrow.DOWN);
-	}
-
-	public void fold()
-	{
-		super.setIcon(Arrow.LEFT);
-	}
-}
