+++ src/main/java/com/larsvansoest/runelite/clueitems/ui/requirement/RequirementStatus.java
@@ -1,0 +1,57 @@
+/*
+ * BSD 2-Clause License
+ *
+ * Copyright (c) 2020, Lars van Soest
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ *    list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package com.larsvansoest.runelite.clueitems.ui.requirement;
+
+import com.larsvansoest.runelite.clueitems.data.EmoteClueItem;
+import java.awt.Color;
+import net.runelite.client.ui.ColorScheme;
+
+/**
+ * Represents values for {@link EmoteClueItem} completion status.
+ *
+ * @author Lars van Soest
+ * @since 2.0.0
+ */
+public enum RequirementStatus
+{
+	Complete(ColorScheme.PROGRESS_COMPLETE_COLOR),
+
+	InProgress(ColorScheme.PROGRESS_INPROGRESS_COLOR),
+
+	InComplete(ColorScheme.LIGHT_GRAY_COLOR),
+
+	Unknown(new Color(254, 254, 254, 69));
+
+	public final Color colour;
+
+	RequirementStatus(Color colour)
+	{
+		this.colour = colour;
+	}
+}
